complete inital restructure
