+++ src/modules/player/components/index.js
@@ -1,9 +1,12 @@
 import { connect } from 'react-redux';
 import PlayerDetails from './main';
 
-import { fetchComments, fetchVideoRating, fetchRelated, fetchDetails } from '../actions';
 import { videosRate } from 'common/interaction/actions';
 import { receiveSetting } from 'common/setting/actions';
+import { fetchDetails,
+         fetchRelated,
+         fetchComments,
+         fetchVideoRating } from '../actions';
 
 const mapStateToProps = ({ playerDetails, setting }) => ({
   playerDetails,
